feat: change styling for main components
