+++ client/src/app.css
@@ -1,7 +1,7 @@
-body {
-  font-family: system-ui, sans-serif;
-  background: #f8f9fa;
-  margin: 0;
-  padding: 2rem;
-  color: #333;
-}
+/*body {*/
+/*  font-family: system-ui, sans-serif;*/
+/*  background: #f8f9fa;*/
+/*  margin: 0;*/
+/*  padding: 2rem;*/
+/*  color: #333;*/
+/*}*/
