+++ client/src/app.tsx
@@ -1,15 +1,18 @@
 import "./app.css";
-import ImageConverter from "./images/ImageConverter.jsx";
-import { useWasm } from "./useWasm.js";
+import ImageConverter from "@/components/images/ImageConverter.jsx";
+import { useWasm } from "./hooks/useWasm.js";
 
 export function App() {
   const { wasmReady, wasmError } = useWasm();
   console.log("WASM ready:", wasmReady);
 
   return (
-    <div>
-      {wasmError && <p>Error initializing WASM: {wasmError.message}</p>}
-      <ImageConverter />
-    </div>
+      <div className="flex flex-col h-screen p-2">
+        {wasmError && <p>Error initializing WASM: {wasmError.message}</p>}
+        <div className="w-full bg-orange-300 flex items-center justify-start text-3xl font-bold rounded-md">
+          <div className='px-6 py-2'>Rustoscope</div>
+        </div>
+         <ImageConverter />
+      </div>
   );
 }
