+++ client/src/components/images/ImageConverter.css
@@ -1,0 +1,99 @@
+/*.image-converter {*/
+/*  max-width: 100%;*/
+/*  margin: 0 auto;*/
+/*  padding: 2rem;*/
+/*  background: white;*/
+/*  border-radius: 1rem;*/
+/*  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);*/
+/*}*/
+
+/*.image-converter-header {*/
+/*  display: flex;*/
+/*  flex-wrap: wrap;*/
+/*  justify-content: space-between;*/
+/*  align-items: center;*/
+/*  gap: 1rem;*/
+/*  margin-bottom: 2rem;*/
+/*}*/
+
+/*.image-converter-header h1 {*/
+/*  font-size: 2rem;*/
+/*  flex: 1 1 auto;*/
+/*  margin: 0;*/
+/*  color: #333;*/
+/*}*/
+
+/*.image-converter-controls {*/
+/*  display: flex;*/
+/*  flex-wrap: wrap;*/
+/*  align-items: center;*/
+/*  gap: 0.75rem;*/
+/*}*/
+
+/*.image-converter-controls input[type="file"] {*/
+/*  display: none;*/
+/*}*/
+
+/*.image-converter-controls button {*/
+/*  padding: 0.5rem 1.25rem;*/
+/*  font-size: 0.95rem;*/
+/*  background-color: #007bff;*/
+/*  border: none;*/
+/*  border-radius: 6px;*/
+/*  color: white;*/
+/*  cursor: pointer;*/
+/*  transition: background-color 0.3s ease;*/
+/*}*/
+
+/*.image-converter-controls button:hover:not(:disabled) {*/
+/*  background-color: #0056b3;*/
+/*}*/
+
+/*.image-converter-controls button:disabled {*/
+/*  background-color: #ccc;*/
+/*  cursor: not-allowed;*/
+/*}*/
+
+/*.conversion-select {*/
+/*  padding: 0.5rem 1.25rem;*/
+/*  font-size: 0.95rem;*/
+/*  border-radius: 6px;*/
+/*  border: 1px solid #ccc;*/
+/*  background-color: white;*/
+/*  color: #333;*/
+/*  cursor: pointer;*/
+/*  transition: border-color 0.3s ease, box-shadow 0.3s ease;*/
+/*}*/
+
+/*.conversion-select:focus {*/
+/*  outline: none;*/
+/*  border-color: #007bff;*/
+/*  box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);*/
+/*}*/
+
+/*.conversion-select:hover {*/
+/*  border-color: #888;*/
+/*}*/
+
+/*.image-preview-container {*/
+/*  display: grid;*/
+/*  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));*/
+/*  gap: 1.5rem;*/
+/*  margin-top: 2rem;*/
+/*  margin-bottom: 2rem;*/
+/*}*/
+
+/*.upload-label {*/
+/*  display: inline-block;*/
+/*  padding: 0.5rem 1.25rem;*/
+/*  background-color: #28a745;*/
+/*  color: white;*/
+/*  border-radius: 6px;*/
+/*  font-size: 0.95rem;*/
+/*  cursor: pointer;*/
+/*  transition: background-color 0.3s ease;*/
+/*}*/
+
+/*.upload-label:hover {*/
+/*  background-color: #218838;*/
+/*}*/
