+++ client/src/components/images/ImagePreview.css
@@ -1,0 +1,34 @@
+/*.image-preview {*/
+/*  width: 100%;*/
+/*  display: flex;*/
+/*  flex-direction: column;*/
+/*  align-items: center;*/
+/*}*/
+
+/*.image-preview h2 {*/
+/*  margin-bottom: 1rem;*/
+/*  font-size: 1rem;*/
+/*  text-align: center;*/
+/*}*/
+
+/*.image-preview img,*/
+/*.empty-preview {*/
+/*  width: 100%;*/
+/*  object-fit: contain;*/
+/*  border-radius: 0.5rem;*/
+/*  background-color: #f9f9f9;*/
+/*  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);*/
+/*}*/
+
+/*.empty-preview {*/
+/*  display: flex;*/
+/*  justify-content: center;*/
+/*  align-items: center;*/
+/*  border: 2px dashed #ddd;*/
+/*  background-color: #f9f9f9;*/
+/*}*/
+
+/*.error-message {*/
+/*  margin-top: 20px;*/
+/*  color: red;*/
+/*}*/
