+++ client/tailwind.config.js
@@ -2,7 +2,7 @@
 export default {
   content: [
       "./index.html",
-      "./src/**/*.{html,js,jsx}"
+      "./src/**/*.{html,js,jsx,ts,tsx}"
   ],
   theme: {
     extend: {},
